refactor: Refactor package structures
