+++ internal/log/ltsv.go
@@ -1,4 +1,4 @@
-package internal
+package log
 
 import (
 	"bufio"
