+++ internal/pattern/node_test.go
@@ -1,4 +1,4 @@
-package internal
+package pattern
 
 import (
 	"fmt"
